scripts/requirements.txt
requests==2.31.0
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0
mysql-connector-python==8.2.0
beautifulsoup4==4.12.2
lxml==4.9.3
scipy==1.11.4
scikit-learn==1.3.2
